{"types":"C","runs":9}

+++ Problem17.java
@@ -2,7 +2,7 @@ import acm.program.ConsoleProgram;
 
 public class Problem17 extends ConsoleProgram {
 	public void run() {
-		double a = readInt("Enter the number: ");
+		double a = readDouble("Enter the number: ");
 		println("The integer part of this number is: " + (int) a + " and the decimal part is: " + ( a - (int) a));
 	}
 }
